# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "myArea REST API",
    "description": ""
  },
  "host": "api.pezentics.com",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/group/addGroup": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Title": {
                  "example": "any"
                },
                "MetaTitle": {
                  "example": "any"
                },
                "Slug": {
                  "example": "any"
                },
                "Summary": {
                  "example": "any"
                },
                "Status": {
                  "example": "any"
                },
                "Profile": {
                  "example": "any"
                },
                "Content": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/addGroupMember": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Status": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "Comment": {
                  "example": "any"
                },
                "MemberType": {
                  "example": "any"
                },
                "GroupId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/updateGroup": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Title": {
                  "example": "any"
                },
                "MetaTitle": {
                  "example": "any"
                },
                "Slug": {
                  "example": "any"
                },
                "Summary": {
                  "example": "any"
                },
                "Status": {
                  "example": "any"
                },
                "Profile": {
                  "example": "any"
                },
                "Content": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "GroupId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/getAllGroup": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/getGroup/{groupId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/getAllGroupMember": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/getGroupMember/{groupMemberId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "groupMemberId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/findAllGroupByDate/{startDate}/{endDate}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/findAllGroupMemberByDate/{startDate}/{endDate}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/deleteGroup/{groupId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/group/deleteGroupMember/{groupMemberId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "groupMemberId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/advert/addadvert": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "AdvertName": {
                  "example": "any"
                },
                "AdvertImgUrl": {
                  "example": "any"
                },
                "AdvertType": {
                  "example": "any"
                },
                "AdvertCaption": {
                  "example": "any"
                },
                "AdvertPage": {
                  "example": "any"
                },
                "AdvertSection": {
                  "example": "any"
                },
                "AdvertRegion": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "IPAddress": {
                  "example": "any"
                },
                "AdvertCost": {
                  "example": "any"
                },
                "Currency": {
                  "example": "any"
                },
                "IsEnabled": {
                  "example": "any"
                },
                "PublishedDateFrom": {
                  "example": "any"
                },
                "PublishedDateFromTo": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/advert/searchAdvert": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "AdvertName": {
                  "example": "any"
                },
                "AdvertType": {
                  "example": "any"
                },
                "AdvertCaption": {
                  "example": "any"
                },
                "AdvertPage": {
                  "example": "any"
                },
                "AdvertSection": {
                  "example": "any"
                },
                "AdvertRegion": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "IPAddress": {
                  "example": "any"
                },
                "AdvertCost": {
                  "example": "any"
                },
                "Currency": {
                  "example": "any"
                },
                "IsEnabled": {
                  "example": "any"
                },
                "PublishedDateFrom": {
                  "example": "any"
                },
                "PublishedDateFromTo": {
                  "example": "any"
                },
                "AdvertId": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/advert/updateAdvert/{advertId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "advertId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "AdvertName": {
                  "example": "any"
                },
                "AdvertType": {
                  "example": "any"
                },
                "AdvertCaption": {
                  "example": "any"
                },
                "AdvertPage": {
                  "example": "any"
                },
                "AdvertSection": {
                  "example": "any"
                },
                "AdvertRegion": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "IPAddress": {
                  "example": "any"
                },
                "AdvertCost": {
                  "example": "any"
                },
                "Currency": {
                  "example": "any"
                },
                "IsEnabled": {
                  "example": "any"
                },
                "PublishedDateFrom": {
                  "example": "any"
                },
                "PublishedDateFromTo": {
                  "example": "any"
                },
                "AdvertId": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/advert/getAdvert/{advertId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "advertId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/advert/getAllAdvert": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/advert/deleteAdvert/{advertId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "advertId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}